scraper/requirements.txt
praw==7.6.0
PyYAML==6.0.0